apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "relay-console-api"
version = "0.1.0"
description = "Relay Console API and worker"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "celery>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.4.0",
  "PyJWT>=2.9.0",
  "psycopg[binary]>=3.2.1",
  "python-slugify>=8.0.4",
  "sqlalchemy>=2.0.35",
  "uvicorn[standard]>=0.30.6"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
